+++ apps/chuni-web/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [next]
+## [3.11.0] - 2026-03-28
 
 - chore: bump deps
 - upgrade to lucide icon v1